Redeclared stuff to stop wngs about signed-vs-unsigned mismatches.
git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@31971 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Include/compile.h
@@ -38,8 +38,8 @@ struct instr {
 };
 
 struct basicblock {
-	size_t b_iused;
-	size_t b_ialloc;
+	int b_iused;
+	int b_ialloc;
 	/* If b_next is non-zero, it is the block id of the next
 	   block reached by normal control flow.
 	   Since a valid b_next must always be > 0,
